WaterlevelLogger - Update commit - Hour 4
Debugging on littlefs script. Current code already has time api implemented, but I make it update every 10 seconds to test the graph so i dont have to wait an hour for the update. Attached a video demo to show graph updating. NOTE TO REVIEWER: Some sections of this code are written with assist by AI tools